# contracts/openapi.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "Voice Platform API",
    "version": "1.0.0-draft",
    "description": "Versioned foundation contract. Feature endpoints are added only with matching ACL and contract tests."
  },
  "paths": {
...
    "/api/v1/direct-messages": {
      "post": {
        "operationId": "openDirectMessage",
        "description": "Authenticated account only. Opens the single canonical direct-message pair when both accounts are active; administrator status never grants access to a third account's DM.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessageOpenRequest" } } }
        },
        "responses": {
          "201": { "description": "New or existing canonical pair", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessage" } } } },
          "400": { "description": "Malformed or self participant", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Other account is absent or blocked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "operationId": "listDirectMessages",
        "description": "Authenticated caller only. Returns only canonical DM pairs containing caller, paired account display name, and caller-local unread_count derived from that caller's monotonic cursor; administrator role grants no exception. A pre-existing pair remains visible to reach retained history after a block.",
        "responses": {
          "200": { "description": "Caller direct-message navigation list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessageList" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/direct-messages/{directMessageID}/messages": {
      "post": {
        "operationId": "sendDirectMessage",
        "description": "Authenticated participant only. The caller must belong to this DM and both participants must be active; administrator role supplies no exception. Reusing the same author-and-DM-scoped client_message_id returns the existing message.",
        "parameters": [
          { "name": "directMessageID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessageMessageCreateRequest" } } }
        },
        "responses": {
          "201": { "description": "Message created or idempotently returned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessageMessage" } } } },
          "400": { "description": "Malformed or oversized message", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "DM is absent, caller is not a participant, or either account is blocked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "operationId": "listDirectMessageHistory",
        "description": "Authenticated DM participant only. Administrator role grants no exception. Returns newest-first cursor pages and retains history for the other participant after a block.",
        "parameters": [
          { "name": "directMessageID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "before", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Newest-first direct-message history page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessageMessagePage" } } } },
          "400": { "description": "Malformed cursor or limit", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "DM is absent or caller is not a participant", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/direct-messages/{directMessageID}/messages/{messageID}": {
      "patch": {
        "operationId": "editDirectMessage",
        "description": "Authenticated DM author only. Pair membership, author ID, non-deleted state, and expected revision are checked atomically; administrator role grants no exception. A conflict intentionally does not disclose which guard failed.",
        "parameters": [
          { "name": "directMessageID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "messageID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessageMessageEditRequest" } } }
        },
        "responses": {
          "200": { "description": "Author message revised", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessageMessageEdited" } } } },
          "400": { "description": "Malformed revision or body", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Message is stale, deleted, unavailable, or not owned by caller; refresh history before retrying", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "operationId": "deleteDirectMessage",
        "description": "Authenticated DM author only. Canonical-pair membership, author ID, and non-deleted state are checked atomically; administrator role grants no exception. The message row, ID, reply reference, and history remain, while body is cleared and a deletion marker with a new revision is written.",
        "parameters": [
          { "name": "directMessageID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "messageID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Author message replaced with a deletion marker" },
          "400": { "description": "Malformed path identifier", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Message is absent, already deleted, unavailable, or not owned by caller", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectMessageOpenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["participant_id"],
        "properties": { "participant_id": { "type": "string", "format": "uuid" } }
      },
      "DirectMessage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "participant_one_id", "participant_two_id", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "participant_one_id": { "type": "string", "format": "uuid" },
          "participant_two_id": { "type": "string", "format": "uuid" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "DirectMessageList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["direct_messages"],
        "properties": {
          "direct_messages": { "type": "array", "items": { "$ref": "#/components/schemas/DirectMessageListItem" } }
        }
      },
      "DirectMessageListItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "other_participant_id", "other_participant_display_name", "created_at", "unread_count"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "other_participant_id": { "type": "string", "format": "uuid" },
          "other_participant_display_name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "unread_count": { "type": "integer", "minimum": 0 }
        }
      },
...
      "DirectMessageMessageCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["client_message_id", "body"],
        "properties": {
          "client_message_id": { "type": "string", "format": "uuid" },
          "reply_to_id": { "type": "string", "format": "uuid", "description": "Must identify a message in this same direct-message pair." },
          "body": { "type": "string", "minLength": 1, "maxLength": 8000 }
        }
      },
      "DirectMessageMessage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "direct_message_id", "author_id", "client_message_id", "body", "revision", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "direct_message_id": { "type": "string", "format": "uuid" },
          "author_id": { "type": "string", "format": "uuid" },
          "client_message_id": { "type": "string", "format": "uuid" },
          "body": { "type": "string" },
          "reply_to_id": { "type": "string", "format": "uuid" },
          "revision": { "type": "integer", "minimum": 1 },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "DirectMessageMessageEditRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["body", "expected_revision"],
        "properties": {
          "body": { "type": "string", "minLength": 1, "maxLength": 8000 },
          "expected_revision": { "type": "integer", "minimum": 1 }
        }
      },
      "DirectMessageMessageEdited": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "direct_message_id", "author_id", "client_message_id", "body", "revision", "created_at", "edited_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "direct_message_id": { "type": "string", "format": "uuid" },
          "author_id": { "type": "string", "format": "uuid" },
          "client_message_id": { "type": "string", "format": "uuid" },
          "body": { "type": "string", "minLength": 1, "maxLength": 8000 },
          "reply_to_id": { "type": "string", "format": "uuid" },
          "revision": { "type": "integer", "minimum": 2 },
          "created_at": { "type": "string", "format": "date-time" },
          "edited_at": { "type": "string", "format": "date-time" }
        }
      },
      "DirectMessageMessagePage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["messages"],
        "properties": {
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/DirectMessageMessageHistoryItem" } },
          "next_cursor": { "type": "string", "format": "uuid" }
        }
      },
...
      "DirectMessageMessageHistoryItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "direct_message_id", "author_id", "client_message_id", "body", "created_at", "revision", "deleted"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "direct_message_id": { "type": "string", "format": "uuid" },
          "author_id": { "type": "string", "format": "uuid" },
          "client_message_id": { "type": "string", "format": "uuid" },
          "body": { "type": "string" },
          "reply_to_id": { "type": "string", "format": "uuid" },
          "reply_preview": { "$ref": "#/components/schemas/DirectMessageReplyPreview" },
          "created_at": { "type": "string", "format": "date-time" },
          "edited_at": { "type": "string", "format": "date-time" },
          "revision": { "type": "integer", "minimum": 1 },
          "deleted": { "type": "boolean" }
        }
      },
      "DirectMessageReplyPreview": {
        "type": "object",
        "description": "Referenced message from the same participant-authorized DM history projection. A soft-deleted source has body:empty and deleted:true.",
        "additionalProperties": false,
        "required": ["id", "author_id", "body", "deleted"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "author_id": { "type": "string", "format": "uuid" },
          "body": { "type": "string" },
          "deleted": { "type": "boolean" }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "request_id"],
        "properties": {
          "code": {
            "type": "string",
            "enum": ["UNAUTHENTICATED", "ACCOUNT_BLOCKED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "VALIDATION_FAILED", "RATE_LIMITED", "MAINTENANCE", "MAINTENANCE_UNAVAILABLE", "ATTACHMENT_TOO_LARGE", "INSUFFICIENT_STORAGE", "ACTIVE_VOICE_LEASE", "VOICE_CHANNEL_UNAVAILABLE", "VOICE_LEASE_UNAVAILABLE", "INTERNAL"]
          },
          "message": { "type": "string" },
          "request_id": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      }
    }
  }
}